refactor: extract result calculation
