+++ src/resultCalculator.ts
@@ -1,0 +1,104 @@
+import { msToTime } from "./helpers/time";
+import { Participant } from "./participant";
+import { Result } from "./result";
+
+export function calculateRankAndSort(results: Result[]): Result[] {
+  results.forEach((result) => (result.result = calculateDifference(result)));
+  sortByTime(results);
+  let rank: number | undefined = 1;
+  for (let i = 0; i < results.length; ++i) {
+    rank = i + 1;
+    if (i !== 0) {
+      results[i].delay = getDelayToPreviousAndFirst(
+        results[0].result,
+        results[i].result,
+        results[i - 1].result
+      );
+
+      const hasSameTime: boolean =
+        (results[i].result as number) - (results[i - 1].result as number) < 100;
+      if (hasSameTime) rank = results[i - 1].rank;
+      results[i].rank = rank;
+    }
+  }
+  results.forEach((x: Result) => {
+    x.startTime = formatTimestampValue(x.startTime);
+    x.finishTime = formatTimestampValue(x.finishTime);
+    x.result =
+      typeof x.result !== "number" || isNaN(x.result)
+        ? x.result
+        : msToTime(x.result);
+  });
+  return results;
+}
+
+function calculateDifference(result: Result): string | number {
+  if (result.startTime === "---") return "DNS";
+  if (result.finishTime === "---") return "DNF";
+  return Date.parse(result.finishTime) - Date.parse(result.startTime);
+}
+
+function formatTimestampValue(value: Date | string): string {
+  if (!value) return "---";
+  return typeof value === "string" ? value : value.toString();
+}
+
+export function mapStartToFinish(
+  starts: Participant[],
+  finishes: Participant[]
+): Result[] {
+  return starts.map((s) => {
+    const res: Result = {
+      rank: 1,
+      numberPlate: s.numberPlate,
+      category: s.category,
+      name: s.name,
+      startTime: (s.time as string) || "---",
+      finishTime:
+        (finishes.find((f) => f.numberPlate == s.numberPlate)
+          ?.time as string) || "---",
+      result: "---",
+      delay: "---",
+    };
+    return res;
+  });
+}
+
+function getDelayToPreviousAndFirst(
+  firstTime: number | string,
+  currentTime: number | string,
+  previousTime: number | string
+): string {
+  if (
+    typeof firstTime !== "number" ||
+    isNaN(firstTime) ||
+    typeof previousTime !== "number" ||
+    isNaN(previousTime) ||
+    typeof currentTime !== "number" ||
+    isNaN(currentTime)
+  )
+    return "---";
+
+  const toFirst = currentTime - firstTime;
+  const toPrevious = currentTime - previousTime;
+  if (typeof currentTime === "number" && !isNaN(currentTime)) {
+    return `+${msToTime(toFirst)} (+${msToTime(toPrevious)})`;
+  }
+  return "---";
+}
+
+/**
+ * 1. smallest time
+ * 2. DNF
+ * 3. DNS
+ */
+function sortByTime(results: Result[]) {
+  results.sort((a, b) => {
+    if (a.result == "DNS" && b.result == "DNF") return 1;
+    if (a.result == "DNF" && b.result == "DNS") return -1;
+    if (!isNaN(a.result as number) && isNaN(b.result as number)) return -1;
+    if (isNaN(a.result as number) && !isNaN(b.result as number)) return 1;
+    if (a.result === b.result) return 0;
+    return (a.result as number) < (b.result as number) ? -1 : 1;
+  });
+}
